+++ SpringBackend/src/main/java/org/falcon/server/controller/TicTacToeRestController.java
@@ -1,4 +1,4 @@
-package org.falcon.TicTacToeServer.controller;
+package org.falcon.server.controller;
 
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
